Rework demo plugin structure
  - Use flat file hierarchy
  - Remove unused/unnecessary cmake definitions
