+++ source/examples/demo-stages/ColorizeStage.h
