+++ source/examples/demo-stages/DemoAntialiasableTriangleStage.h
