+++ source/examples/demo-stages/DemoAntialiasingPipeline.h
