+++ source/examples/demo-stages/DemoDOFCubeStage.h
