+++ source/examples/demo-stages/DemoDrawableStage.h
