+++ source/examples/demo-stages/DemoMultiFramePipeline.h
