+++ source/examples/demo-stages/DemoRenderStage.h
