+++ source/examples/demo-stages/DemoSSAOPipeline.h
