+++ source/examples/demo-stages/DemoSSAOPostprocessingStage.h
