+++ source/examples/demo-stages/DemoSSAORenderingStage.h
