+++ source/examples/demo-stages/DemoTransparencyPipeline.h
