+++ source/examples/demo-stages/DemoTransparencyStage.h
