+++ source/examples/demo-stages/LightTestPipeline.h
