+++ source/examples/demo-stages/LightTestStage.h
