+++ source/examples/demo-stages/SpinningRectStage.h
