+++ source/examples/demo-stages/plugin.cpp
@@ -1,25 +1,25 @@
 
 #include <cppexpose/plugin/plugin_api.h>
 
-#include <demo-stages/ColorizeStage.h>
-#include <demo-stages/DemoAntialiasableTriangleStage.h>
-#include <demo-stages/DemoAntialiasingPipeline.h>
-#include <demo-stages/DemoDOFCubeStage.h>
-#include <demo-stages/DemoDOFPipeline.h>
-#include <demo-stages/DemoMultiFramePipeline.h>
-#include <demo-stages/DemoPipeline.h>
-#include <demo-stages/DemoRenderStage.h>
-#include <demo-stages/DemoSSAOPipeline.h>
-#include <demo-stages/DemoSSAOPostprocessingStage.h>
-#include <demo-stages/DemoSSAORenderingStage.h>
-#include <demo-stages/DemoStage.h>
-#include <demo-stages/DemoTransparencyPipeline.h>
-#include <demo-stages/DemoTransparencyStage.h>
-#include <demo-stages/LightTestPipeline.h>
-#include <demo-stages/LightTestStage.h>
-#include <demo-stages/ShaderDemoPipeline.h>
-#include <demo-stages/SpinningRectStage.h>
-#include <demo-stages/TimerStage.h>
+#include "ColorizeStage.h"
+#include "DemoAntialiasableTriangleStage.h"
+#include "DemoAntialiasingPipeline.h"
+#include "DemoDOFCubeStage.h"
+#include "DemoDOFPipeline.h"
+#include "DemoMultiFramePipeline.h"
+#include "DemoPipeline.h"
+#include "DemoRenderStage.h"
+#include "DemoSSAOPipeline.h"
+#include "DemoSSAOPostprocessingStage.h"
+#include "DemoSSAORenderingStage.h"
+#include "DemoStage.h"
+#include "DemoTransparencyPipeline.h"
+#include "DemoTransparencyStage.h"
+#include "LightTestPipeline.h"
+#include "LightTestStage.h"
+#include "ShaderDemoPipeline.h"
+#include "SpinningRectStage.h"
+#include "TimerStage.h"
 
 
 CPPEXPOSE_PLUGIN_LIBRARY
